# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "engine.Template": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "spec": {
                        "$ref": "#/components/schemas/engine.CreateVMParams"
                    }
                },
                "required": [
                    "description",
                    "id",
                    "name",
                    "spec"
                ],
                "type": "object"
            },
...
            "jobs.Job": {
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    },
                    "finished_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "logs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "result": {
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string"
                    },
                    "state": {
                        "$ref": "#/components/schemas/jobs.State"
                    },
                    "target": {
                        "type": "string"
                    }
                },
                "required": [
                    "action",
                    "created_at",
                    "id",
                    "label",
                    "logs",
                    "state",
                    "target"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/jobs/{id}/wait": {
            "get": {
                "description": "Wait for one job ID until completion through a finite server-sent event stream. The done event returns the final job. Failed jobs are errors; connection loss does not cancel or prove failure of the underlying job. Inspect getJob or wait again with the same ID.",
                "operationId": "waitJob",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "503": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "waitJob",
                "tags": [
                    "jobs"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "stream": {
                        "event": "done",
                        "failureValue": "failed",
                        "stateField": "state"
                    },
                    "transport": "sse"
                }
            }
        },
...
        "/api/templates": {
            "get": {
                "description": "List saved machine-profile templates and their IDs. Each template carries guest provisioning (cloud-init or Ignition) and hardware defaults, used to pre-fill VM creation.",
                "operationId": "listTemplates",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/engine.Template"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listTemplates",
                "tags": [
                    "templates"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "transport": "json"
                }
            },
            "post": {
                "description": "Save a reusable machine-profile template. The spec mirrors the createVM request (guest provisioning plus hardware) minus the VM name. Requires an administrator account.",
                "operationId": "createTemplate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.TemplateRequest",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/engine.Template"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "createTemplate",
                "tags": [
                    "templates"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json"
                }
            }
        },
        "/api/templates/{id}": {
            "delete": {
                "description": "Delete a saved machine-profile template by ID from listTemplates. Requires an administrator account.",
                "operationId": "deleteTemplate",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "deleteTemplate",
                "tags": [
                    "templates"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json"
                }
            },
            "get": {
                "description": "Read a single machine-profile template by ID from listTemplates. The returned spec is a full createVM request without a name, so it can be edited and passed to createVM to customize fields that instantiateTemplate does not override, such as network, addresses, interfaces, disks, or guest_setup.",
                "operationId": "getTemplate",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/engine.Template"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "getTemplate",
                "tags": [
                    "templates"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "transport": "json"
                }
            },
            "put": {
                "description": "Replace a saved template by ID from listTemplates with the supplied name, description, and spec. Requires an administrator account.",
                "operationId": "updateTemplate",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.TemplateRequest",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/engine.Template"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "updateTemplate",
                "tags": [
                    "templates"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json"
                }
            }
        },
        "/api/templates/{id}/instantiate": {
            "post": {
                "description": "Create a VM from a saved template by ID from listTemplates. This is a convenience path that keeps the template spec as-is and only lets you set a unique VM name and optional cpus, memory_mib, disk_size_gib, or autostart overrides. To change anything else, including networking, addresses, interfaces, disks, or guest provisioning, do not use this endpoint: read the template with getTemplate, edit its spec (a full createVM body without name), and pass it to createVM. Returns a queued job like createVM; the job result contains the new VM ID. Requires an administrator account.",
                "operationId": "instantiateTemplate",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/engine.TemplateInstanceParams",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "instantiateTemplate",
                "tags": [
                    "templates"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json",
                    "wait": {
                        "idField": "id",
                        "operationId": "waitJob",
                        "parameter": "id"
                    }
                }
            }
        },
...
        "/api/vms": {
            "get": {
                "description": "List all VMs with their manifest, current phase, process ID, and boot time. Use manifest.id as the id argument for VM operations. Read-only; does not start VMs.",
                "operationId": "listVMs",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/engine.VMView"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listVMs",
                "tags": [
                    "vms"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "transport": "json"
                }
            },
            "post": {
                "description": "Create a VM and its disks. Discover image keys and provisioning capabilities with listCatalog, existing media with listMedia, and network IDs with listNetworks first. Body is an API creation request, not a persisted manifest. Provide name, cpus, memory_mib (MiB), disk_size_gib (GiB), and an image key or installation ISO. Example body: {\"name\":\"agent-vm\",\"cpus\":2,\"memory_mib\":2048,\"disk_size_gib\":20,\"username\":\"maco\",\"image\":\"<key from listCatalog>\"}. The job result contains the new VM ID; creation does not imply the VM is running. Use startVM afterward when needed. Requires an administrator account.",
                "operationId": "createVM",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/engine.CreateVMParams",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "createVM",
                "tags": [
                    "vms"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json",
                    "wait": {
                        "idField": "id",
                        "operationId": "waitJob",
                        "parameter": "id"
                    }
                }
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ],
...
}